+++ frontend/src/pages/StatusPages/modals/CreateGroupModal.tsx
@@ -4,22 +4,19 @@ import Dialog from '@/components/Dialog/Dialog'
 import useApiService from '@/hooks/useApi'
 import { IStatusPage } from '@/types/models'
 
-import GroupForm, { FormValues } from './components/GroupForm'
-
-import { GroupValue } from './types'
+import GroupForm, { FormValues } from '../components/GroupForm'
 
 interface Props {
   onClose: () => void
-  group: GroupValue
   statusPage: IStatusPage
 }
 
-const EditGroupModal: React.FC<Props> = ({ onClose, group, statusPage }) => {
+const CreateStatusPageGroupModal: React.FC<Props> = ({ onClose, statusPage }) => {
   const { apiService } = useApiService()
   const queryClient = useQueryClient()
 
-  const createStatusPageMutation = useMutation({
-    mutationFn: (value: FormValues) => apiService.editStatusPageGroup(statusPage.id, group.id, value),
+  const mutation = useMutation({
+    mutationFn: (value: FormValues) => apiService.createStatusPageGroup(statusPage.id, value),
     onSuccess: () => {
       onClose()
       queryClient.invalidateQueries({ queryKey: ['get-status-page', statusPage.id] })
@@ -27,16 +24,16 @@ const EditGroupModal: React.FC<Props> = ({ onClose, group, statusPage }) => {
   })
 
   const handleSubmit = (values: FormValues) => {
-    createStatusPageMutation.mutateAsync({
+    mutation.mutateAsync({
       ...values
     })
   }
 
   return (
-    <Dialog onClose={onClose} title="Edit group">
-      <GroupForm onSubmit={handleSubmit} group={group} />
+    <Dialog onClose={onClose} title="Create group">
+      <GroupForm onSubmit={handleSubmit} />
     </Dialog>
   )
 }
 
-export default EditGroupModal
+export default CreateStatusPageGroupModal
