+++ frontend/src/pages/StatusPages/modals/EditComponentModal.tsx
@@ -4,19 +4,22 @@ import Dialog from '@/components/Dialog/Dialog'
 import useApiService from '@/hooks/useApi'
 import { IStatusPage } from '@/types/models'
 
-import ComponentForm, { FormValues } from './components/ComponentForm'
+import ComponentForm, { FormValues } from '../components/ComponentForm'
+
+import { ComponentValue } from '../types'
 
 interface Props {
   onClose: () => void
+  component: ComponentValue
   statusPage: IStatusPage
 }
 
-const CreateComponentModal: React.FC<Props> = ({ onClose, statusPage }) => {
+const EditComponentModal: React.FC<Props> = ({ onClose, component, statusPage }) => {
   const { apiService } = useApiService()
   const queryClient = useQueryClient()
 
   const mutation = useMutation({
-    mutationFn: (value: FormValues) => apiService.createStatusPageComponent(statusPage.id, value),
+    mutationFn: (value: FormValues) => apiService.editStatusPageComponent(statusPage.id, component.id, value),
     onSuccess: () => {
       onClose()
       queryClient.invalidateQueries({ queryKey: ['get-status-page', statusPage.id] })
@@ -30,10 +33,10 @@ const CreateComponentModal: React.FC<Props> = ({ onClose, statusPage }) => {
   }
 
   return (
-    <Dialog onClose={onClose} title="Create component">
-      <ComponentForm onSubmit={handleSubmit} />
+    <Dialog onClose={onClose} title="Edit component">
+      <ComponentForm onSubmit={handleSubmit} component={component} />
     </Dialog>
   )
 }
 
-export default CreateComponentModal
+export default EditComponentModal
